Small fixes (get rid of static analyzer's warnings)

+++ Source/Common/CmdLine.h
@@ -30,13 +30,13 @@ class CCmdLine
         CCmdLine &operator=(const CCmdLine& p);
         explicit CCmdLine(LPCTSTR szCmdLine);
         void Parse(LPCTSTR szCmdLine);
-        int AddParam(LPCTSTR szParam);
+        size_t AddParam(LPCTSTR szParam);
 
         //Returns command line without module name
         CString OnlyParams();
-        CString operator[](int nIndex);
+        CString operator[](size_t nIndex);
         CString ModuleName();
-        bool GetNextFile(CString& FileName, int& nIndex);
+        bool GetNextFile(CString& FileName, size_t& nIndex);
         bool IsOption(LPCTSTR Option, bool bUsePrefix = true);
         size_t GetCount();
     private:
